relegate Wanda++ sparsifier to prototype - Since WandaPlusPlusSparsifier is still in prototype stage, we move it to the prototype directory.
update annotations and remove rebudnat error handler

update indexing logics for Wanda++ using Wanda
- Since both paper uses same index logics (extraction, sorting, and updating), Wanda++ can use the same logic as Wanda.
- For using it, `_apply_sparsity_pattern` is moduled to be used in both Wanda and Wanda++.

remove hardware-specific test case
- This test case is not fully implemented in `wanda_pp.py`
