+++ torchao/prototype/sparsity/pruner/__init__.py
@@ -14,4 +14,5 @@ __all__ = [
     "BiasHook",
     "FakeStructuredSparsity",
     "SaliencyPruner",
+    "WandaPlusPlusSparsifier",
 ]
